bump chart - refactor code into module
