+++ bump_chart/models/bump_chart_model.py
