+++ bump_chart/views/bump_chart_view.py
@@ -5,7 +5,7 @@ import seaborn
 import textwrap
 
 
-from bump_chart_model import BumpChartModel
+from bump_chart.models.bump_chart_model import BumpChartModel
 from typing import Tuple
