+++ generate_bump_chart.py
@@ -1,0 +1,4 @@
+from bump_chart import generate_bump_chart
+
+if __name__ == "__main__":
+    generate_bump_chart()
